Move contexts to src/client/contexts
