+++ src/client/components/App/index.tsx
@@ -2,7 +2,7 @@ import { FunctionComponent, Fragment, h } from 'preact';
 import Books from '~client/components/Books';
 
 import '~client/components/App/styles.css';
-import { OptionsProvider } from '~client/components/contexts/Options';
+import { OptionsProvider } from '~client/contexts/Options';
 import Nav from '~client/components/Nav';
 
 const App: FunctionComponent = () => (
